requirements.txt
google-cloud-dialogflow==2.21.0
python-dotenv==1.0.0
python-telegram-bot==13.15